rename proptest_arbitrary
